Add padding for code samples.
On some code examples - the "Copy" button covers part of the code so you can't see all of the code example. I've added padding to the code tag so the code can always be seen. 

I've uploaded a screenshot to tinypic as an example of the issue: http://i65.tinypic.com/8vu1lc.png