fix: ensure Sass/Js default values to be escaped
Despite the expected Handlebars behavior (escaping everything "SafeString" is used), the HTML that may be in Sass and Js  values are not escaped. No option is used (like "noEscape") that would explain this behavior.

This commit add the "escapeHTML" helper and ensure that Sass and Js values are always escaped.